fix: vsphere image repo
The old repository has been deleted. This means the driver is currently
not possible to deploy, and even currently running clusters fail after a
restart as the DeameonSet has an ImagePullPolicy of Always. See:
https://github.com/kubernetes-sigs/vsphere-csi-driver/issues/3053

As of writing, only 3.1.2 to 3.3.1 is available in this registry. This
"officially" supports Kubernetes 1.26 to 1.30. Since older drivers are
not available, I removed some feature-gating for those unavailable
versions.